Pull out table to another new component
