+++ Tasker.WebApp/ClientApp/src/components/TodoForm.js
@@ -7,10 +7,10 @@ const HEADER = {
     'Content-Type': 'application/json'
 };
 
-const URL = "http://localhost:5000/api/todos/"
-
+const URL = "http://localhost:5000/api/todos/";
 
 export class TodoForm extends React.Component {
+
     constructor(props) {
         super(props);
         this.state = { title: '', description: '', deadline: '', priority: 1, state: '', date: null };
@@ -27,14 +27,13 @@ export class TodoForm extends React.Component {
 
     async handleSubmit(event) {
         event.preventDefault();
-        var data = JSON.stringify({
+        const data = JSON.stringify({
             "title": event.target.title.value,
             "description": event.target.description.value,
             "deadline": this.state.date,
             "priority": this.state.priority,
             "state": 1
         });
-        console.log(data);
         await fetch(URL, {
             headers: HEADER,
             method: 'POST',
